LayoutTests/media/video-display-aspect-ratio.html

<!DOCTYPE html>
<title>Test video dimension aspect ratio.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    assert_equals(video.videoWidth, 0);
    assert_equals(video.videoHeight, 0);

    video.onloadedmetadata = t.step_func_done(function() {
        assert_equals(video.videoWidth, 426);
        assert_equals(video.videoHeight, 240);
    });

    video.src = "content/test-par-16-9.mp4";
});
</script>
